requirements.txt
Flask
pytesseract
pillow
flask-restx
